refactor(mpm): split the new verification cases into two sibling modules
The combined file reached 2237 lines against the repository's 1200-line
module-size budget, which is a gate. limit_states.py takes the Rankine
plastic-limit case and order_of_accuracy.py takes the manufactured
solution, its patch-test companion and the temporal refinement;
verification.py keeps the shipped four checks and points at both.

The dependency runs one way: order_of_accuracy imports the two wall
configurations and the two set-up helpers from verification rather than
growing a second copy of the 'wall bands land on the column' knowledge,
and verification imports nothing back.
